refactor: rename index.ds.ts -> index.d.ts, d for declaractions
